+++ config.js
@@ -39,7 +39,8 @@ exports.cleverService = {
   discoveryTimeoutMs: 4000,
   retryIntervalMs: 8000,
   maxRetryIntervalMs: 30000,
-  fallbackAddresses: ['127.0.0.1']
+  fallbackAddresses: ['127.0.0.1'],
+  httpPorts: [80, 8000]
 }
 
 // Appearance settings